==> 0_junta_censo.ipynb <==
censo_arquivos = ["data/dados_censo/MICRODADOS_CADASTRO_CURSOS_2012.CSV",
"data/dados_censo/MICRODADOS_CADASTRO_CURSOS_2013.CSV",
"data/dados_censo/MICRODADOS_CADASTRO_CURSOS_2014.CSV",
"data/dados_censo/MICRODADOS_CADASTRO_CURSOS_2015.CSV",
"data/dados_censo/MICRODADOS_CADASTRO_CURSOS_2016.CSV",
"data/dados_censo/MICRODADOS_CADASTRO_CURSOS_2017.CSV",
"data/dados_censo/MICRODADOS_CADASTRO_CURSOS_2018.CSV",
"data/dados_censo/MICRODADOS_CADASTRO_CURSOS_2019.CSV",
"data/dados_censo/MICRODADOS_CADASTRO_CURSOS_2020.CSV",
"data/dados_censo/MICRODADOS_CADASTRO_CURSOS_2021.CSV",
"data/dados_censo/MICRODADOS_CADASTRO_CURSOS_2022.CSV",
"data/dados_censo/MICRODADOS_CADASTRO_CURSOS_2023.CSV"]

import pandas as pd

all_dfs = []
for file_name in censo_arquivos:
    # Create the full file path
    # file_path = os.path.join(data_dir, file_name)
    
    # Read the Excel file with the same parameters you used
    df = pd.read_csv(file_name, sep=';', encoding='latin1')
    # df = df.iloc[:-1]
    
    # Optional: Add a column to identify which file this data came from
    df['source_file'] = file_name
    
    # Append to our list of dataframes
    all_dfs.append(df)
    
    # Print confirmation (optional)
    print(f"Loaded {file_name}, shape: {df.shape}")

# Combine all dataframes into one
raw_df = pd.concat(all_dfs, ignore_index=True)

# raw_df = raw_df.dropna()

raw_df.to_parquet("data/dados_censo/MICRODADOS_CADASTRO_CURSOS_2012_2023.parquet")

del raw_df

df = pd.read_parquet("data/dados_censo/MICRODADOS_CADASTRO_CURSOS_2012_2023.parquet")

df.head()
