==> 0_junta_trajetorias.ipynb <==
arquivos = ["data/indicadores_trajetoria/indicadores_trajetoria_educacao_superior_2011_2020.xlsx",
"data/indicadores_trajetoria/indicadores_trajetoria_educacao_superior_2012_2021.xlsx",
# "data/indicadores_trajetoria/indicadores_trajetoria_educacao_superior_2013_2021.xlsx",
"data/indicadores_trajetoria/indicadores_trajetoria_educacao_superior_2013_2022.xlsx",
"data/indicadores_trajetoria/indicadores_trajetoria_educacao_superior_2014_2023.xlsx",
"data/indicadores_trajetoria/indicadores_trajetoria_educacao_superior_2015_2023.xlsx",
"data/indicadores_trajetoria/indicadores_trajetoria_educacao_superior_2016_2023.xlsx",
"data/indicadores_trajetoria/indicadores_trajetoria_educacao_superior_2017_2023.xlsx",
"data/indicadores_trajetoria/indicadores_trajetoria_educacao_superior_2018_2023.xlsx",
"data/indicadores_trajetoria/indicadores_trajetoria_educacao_superior_2019_2023.xlsx"]

import pandas as pd

all_dfs = []
for file_name in arquivos:
    # Create the full file path
    # file_path = os.path.join(data_dir, file_name)
    
    # Read the Excel file with the same parameters you used
    df = pd.read_excel(file_name, header=[8])
    df = df.iloc[:-1]
    
    # Optional: Add a column to identify which file this data came from
    df['source_file'] = file_name
    
    # Append to our list of dataframes
    all_dfs.append(df)
    
    # Print confirmation (optional)
    print(f"Loaded {file_name}, shape: {df.shape}")

# Combine all dataframes into one
raw_df = pd.concat(all_dfs, ignore_index=True)

raw_df['NU_ANO_INGRESSO'].unique()

raw_df

combined_df = raw_df.dropna()

combined_df

new_columns = [
    "Código da Instituição", "Nome da Instituição", "Categoria Administrativa", "Organização Acadêmica",
    "Código do Curso de Graduação", "Nome do Curso de Graduação", "Código da Região Geográfica do Curso",
    "Código da Unidade Federativa do Curso", "Código do Município do Curso", "Grau Acadêmico", 
    "Modalidade de Ensino", "Código da área do Curso segundo a classificação CINE BRASIL", 
    "Nome da área do Curso segundo a classificação CINE BRASIL", 
    "Código da Grande Área do Curso segundo a classificação CINE BRASIL", 
    "Nome da Grande Área do Curso segundo a classificação CINE BRASIL", "Ano de Ingresso", "Ano de Referência",
    "Prazo de Integralização em Anos", "Ano de Integralização do Curso", 
    "Prazo de Acompanhamento do Curso em anos", "Ano Máximo de Acompanhamento do Curso", 
    "Quantidade de Ingressantes no Curso", "Quantidade de Permanência no Curso no ano de referência", 
    "Quantidade de Concluintes no Curso no ano de referência", 
    "Quantidade de Desistência no Curso no ano de referência", 
    "Quantidade de Falecimentos no Curso no ano de referência", "Taxa de Permanência - TAP", 
    "Taxa de Conclusão Acumulada - TCA", "Taxa de Desistência Acumulada - TDA", 
    "Taxa de Conclusão Anual - TCAN", "Taxa de Desistência Anual - TADA", "source_file"
]
combined_df.columns = new_columns

combined_df.to_parquet("data/indicadores_trajetoria/indicadores_trajetoria_educacao_superior_all.parquet")
